Initial commit + Start rewriting entire API to use the new concurrency model
Swift 5.5 is finally here, and with it comes the new concurrency model, that supports calling asynchronous functions in a much easier way.

This commit starts the process of rewriting the entire backend of [Starlight](https://github.com/hyperspacedev/starlight), adapting it to the new async/await syntax, ditching completion handlers (or at least partly).